Network window: Prevent device capture from blocking inputs on title screen
When captured, pads get marked as "in use" in the corresponding
subsystem (the wrapper around Windows raw inputs or the wrapper around
XInput). Capturing ignores in-use pads, otherwise players could steal
input bindings from each other. The titlescreen uses capturing as a
convenience method to detect the first device to press a button. Since
the network window runs first, if the network window is visible on the
title screen, the network window captures the device and marks it as in
use, preventing the same device from being used to activate the title
screen.