Events: Annotate "foreground" concepts, replace instance tracking
A lot of event-related code relied on the fact that many times, only on
event instance per event type is created at a time. This is still true,
so the various `*::instance` properties did still work, but they were
proof that I didn't have complete analysis of the event infrastructure.

Instead, after quite a bit of reversing, I understand quite a bit more,
and this commit takes advantage of those new learnings. The two critical
pieces are:

* Often, instances may be _constructed_ but not in active use. For example,
  `VsCharaSelect` begins initializing the next `VsStageSelect` event inside
  its own `RegisterTasks` implementation. This led to some interesting
  issues with the main menu, for example.
* Cancelling events in various states is possible, easy, and commonly done.
  For example, `VsCharaSelect` cancels the potentially already-initialized
  `VsStageSelect` event if the user backs out to the main menu.

Expanding on the full lifecycle of an event, events are:
* Constructed,
* initializing,
* in the foreground (ex. `RegisterTasks` was called),
* moved to the background (or otherwise considered "stopped"),
* finalizing (ex. releasing system resources),
* and finally destroyed.

Combining these, and with a little knowledge of the root event, we can
traverse the full hierarchy of "foreground" events, and can easily execute
arbitrary code when a certain hierarchy exists. For example, the overlay
can output `VsCharaSelect` data without having to track instances- it can
just traverse each foreground event in the event hierarchy formed by the
nested `EventBaseWithEC` instances to determine if the lowest event is
a `VsCharaSelect` event.

This also means "skip" code is trivial to write- instead of hijacking
`QueueEvent` calls, just let the game schedule whatever, then cancel
it after. It also led to learning a lot more about the "event listeners"
infrastructure- more on both likely to follow after this commit.