Sidecar sessions: Break battle conditions into separate messages
Eventually, it'd be nice to integrate with the existing character and
stage selection screens. In order for that to really be valuable, we
need more granular messages, but those messages don't come without
tradeoffs.

The main negative is that the protocol becomes significantly more tied
to the game's UI and that consequently the number of messages in the
protocol goes way up to reflect every UI element and interaction. I've
tried for literally years to come up with a UI-agnostic session system,
but every system I've come up with is easily 2 or 3x the amount of code
compared to just writing the protocol, and is complicated enough that
both the library and calling code would  likely have bugs anyway. I
could've written several game-specific implementations in the time it
took me to think of and then eventually reject a general solution-
might as well just ship a game-specific impementation, instead of
wasting my own time.

However, once you assume your protocol is game-specific, a lot of other
pieces fall into place. For example, you don't have to assume the
client could receive any message in any scene, so instead of having a
"ready" _state_ you have to ship over the wire, you can just have a
one-time ready _message_. Since the message is a one-time message, you
don't have to worry about lobby members joining late receiving the state
and accidentally transitioning when they shouldn't. It also means that
synchronization messages in the protocol can now refer to scenes in the
specific game and UI. For example, the protocol can include a "scene
loaded" message noting that the client is displaying the character
select screen. That alone could be represented as an integer in a
generic protocol, but exposing the scene to the caller of the session
client means that the client can rely on the caller to referee transitions
between scenes (ex. should I wait for everyone, or just continue). It's
easy to accidentally write a bug in the caller code that doesn't handle
transitions correctly, so this API is loaded with footguns, but I'm not
sure the alternative of an incomprehensible, high-LOC abstraction would
have fewer footguns.