Overlay: Move into sf4e target
I've kept the overlay in its own CMake target for a long time as an
ideal performance optimization, because changes to the overlay shouldn't
necessitate relinking the rest of sf4e and vice versa. That constraint has
bordered on unhelpful for a while- not being able to communicate between
the "app" code and the "overlay" code results in the overlay library
storing its own error messages, for example. It's also been leaky the
whole time, because there's at minimum a header dependency from the existing
sf4e target to the existing overlay target for initialization that wasn't
reflected in the CMake targets.

Instead, just include the overlay code with the sf4e library, which is
mostly just a library for want of a compilation target. The performance
hit isn't that significant anyway- this isn't a full engine build. This
also makes it way easier to put session code in a more helpful place-
future session callbacks likely shouldn't live in the overlay.