Session: Use explicit "all ready" message instead of deriving ready
As 7f7f92af points out, having a ready "state" sent over the wire can be
hard to get right. In sf4e's case, because the ready request numbers
weren't being reset the moment the state was recieved, clients joining
a session where both other clients in the session were already ready
would immediately be placed in the game, despite the GGPO session being
started. This could be avoided by sending _exactly_ one session-data-update
message with the two ready requests set, but then also requires that the
client expose every intermediary state instead of just the state at the
end of each `Step()` call.

Instead, just use the server as a source of truth, with a one-shot
all-ready message to advance the users to the next scene.