Session: Sync battle start outside GGPO for better UI feedback
GGPO is acting as a synchronization mechanism, but since the GGPO
session's blocking only comes into play once the game is simulating
frames, the two players in the game will get like 8 frames into the game
and the freeze while other delayed members load in. Instead, sync
outside GGPO so that we can show the "now loading" screen.

Kinda cursed implementation, but all of this is cursed at this point.