Session: assign (reusable) connection IDs to each connected client
I've been wanting to avoid ID-based implementations for quite some time,
as I have a pet theory that it's led to some pretty bad bugs in several
video games. That being said, if I ever want to get the actual queueing
implementation out of the server (and I do so that a central server
could be hosted), I kind of have to implement IDs- without them, there's
no way for anything but the server to refer to a specific client.

At first I was thinking about just incrementing an integer and assigning
successive increments as IDs to clients. This would work fine up to a
point, but you'd either need a strategy to handle integer overflow, or
use an ID large enough that overflow will never happen for the whole
lifetime of the process. A 32-bit ID counter could believably overflow-
4 billion connections for one process launch is something that has
happened before- so it would need a map to prevent reusing an ID that
was already used after overflow. However, even with overflow solved,
that puts the spotlight on a different form of scale: it'd support a
large number of successive connections over time, but would be unrelated
to the maximum number of concurrent users. Since session information is
intended to be used in a near-real-time context, we need server push,
which means we need to hold open connections. Holding those connections
open and running the application logic to manipulate them would probably
become the bottleneck, not by the ID strategy or any data store backing
it. Scaling up concurrent connections is really only achievable by
minimizing overhead per connection and providing more hosts for clients
to connect to, and the first will only go so far. That means eventually
you may need two server hosts, but since that incrementing ID would only
be valid on the server host that issued it, server hosts wouldn't be able
to reference identities issued by each other. Effectively, scaling would
create two distinct sets of users in a "lobby 1" and "lobby 2" scenario-
not ideal. So how can we get temporary globally unique identifiers that
we can share against servers in the cluster?

Using random self-generated identifiers like 4K self-generated certs as
connection IDs would be a practical approach and observably work
fine for something like the Kademlia DHT, but I honestly just don't
trust myself to write that correctly.

Just using the IP address of the remote host opening the connection to
the server as an identity shareable between server hosts in a cluster
would _almost_ work... if it weren't for NAT. This is actually an issue
with the use of public IP addresss to set up GGPO connections in sf4e
today, too. Consider what would happen if two clients connect to two
different servers in a cluster. If those two clients are behind the
same address-dependent NAT (aka "type 2"), the NAT router could overwrite
the source IP of both clients to _the same source IP and port_ when
sending to the server, and map back to the client when receving a packet
from the server based on the server source address. In other words, that
address is still only valid when used in outbound packets from the server
that is managing the connection: if server B sends a packet to a.b.c.d
because server A informed it that a.b.c.d exists, server B may end up
actually getting its packets delivered to the wrong host. There's a
straightforward solution to that- if server B sent the message to server
A to forward on to the client, server B could still use an ID generated
by server A. At that point, server B would only need to be able to
distinguish between IDs generated by itself and those generated by
server A.

...I reinvented prefix routing.

Prefix routing is basically the solution I ended up going with: each
session server can delegate identities prefixed with its own identity,
effectively becoming a router in an overlay network. Each server's
identity can be an arbitrary string as long as it's unique and routable-
an IP, a hostname, whatever- which is much easier to assert in the
context of running a cluster than about the clients connected to the
cluster. In a way, you can almost think of this as a domain-specific
sharded database, where the host that needs the database row will
always be the one handling the connection to the client.

In addition to moving queue logic from the server to the leader
of the session allowing for a limited form of clustering, it opens
up convenient client-to-client relaying (ex. to exchange ICE candidates).
As an optimization, instead of using an incrementing ID and a map on
each host, we can instead just use the socket handles themselves as IDs:
the OS will only give us socket handles that aren't in use, and all the
querying we need to do is entirely predicated on getting a socket handle
anyway. When the connection closes, the socket handle is necessarily
dead- we barely even need to clean up. This means it's simple enough
that we don't even need a separate DB process or even embedded sqlite.
It also can dovetail with but does not require any account
authentication.

It's also _incredibly_ ridiculously overengineered, and I can't imagine
anyone else wanting to ever use it. At a large enough scale and
particularly once you introduce authentication, this almost certainly
just turns into a Valkey cluster referenced by the connection hosts, but
a session service that allows unauthenticated operation is very nice to
have if I need it. In particular, it means any community running the
software doesn't really have a central chokepoint- with no need to set
up any authentication, "hosting" is as simple as just running the
binary. No Docker, no additional process management or daemons, nothing-
just the binary.