Sessions: Implement message forwarding/relaying between clients
With bc36849c in place and each client having an ID associated with it,
we can now really easily implement forwarding. This implementation
doesn't take into account the clustering bc36849c suggests yet, but this
is the skeleton we'll need to both use the forwarding as a way to move
the queueing and sequencing logic out to the clients as well as to use
the server as a signaling service for the GameNetworkingSockets
implementation.

With this as a signaling service, we need to get GGPO tunneled through
GameNetworkingSockets, and then we don't need any port forwarding from
anyone except the host, but then I have to hack GGPO again, and uuuugh.