Session: Expand connection IDs to include separate host/user fields
I originally was doing _literal_ prefix routing, and I probably will
return to that at some point, but for now it's helpful to be able to
easily reference just the portion of the CID that refers to the host
identity and just the portion of the CID that refers to the "user"
(the socket) without having to worry about manually splitting on a
delimiter or something. This does eventually likely just look like
a URL, but the first URL parser libraries I went to are C++20 only,
which this project very much is not.