Session: Add lobby IDs and the "null" lobby to define pre-connection behavior
I've been playing fast and loose with the `LobbyData` class basically
ever since I implemented it as a way of just getting two clients to
connect, assuming that basically every implementation will check
`_connected` first. Instead, just add a "null" lobby that can never
contain members. In order to distinguish the null lobby from other
lobbies, it's convenient to introduce the concept of lobby IDs, parallel
to connection IDs. Clients will start with a view of the null lobby,
even though they can't participate in it, to ensure that we have a
consistent starting point for every client. After connecting, the local
`lobbyData` will be populated, and after disconnecting, the local
`lobbyData` will be reset to the null lobby. This has an unintuitively
helpful consequence: it means we can instead treat `lobbyData` more as
"the view of an arbitrary lobby" instead of the one the client is
actually participating in, which could open up the ability to do things
like switch more transparently between lobbies in the future.

This is eventual groundwork for lobby "ownership" as well. Once we get
clients respecting the concept of the null lobby, we can implement it
in the server, and only create the lobby on request. On creation, we can
grant ownership to the CID that requested the lobby, which will further
enable moving stuff like the battle signaling to the clients.