Session: Add required hello message before joining
I've been thinking about hypothetical ways to extend the server, top of
mind of which is definitely authorization-related. Having authorization
hints sent to the client would make it easier to disambiguate which
servers authorize unauthenticated connections and which servers require
additional authorization, but that requires messages before the join
message. Eventually, we probably need to support this anyway, but
unfortunately it increases the number of round trips we need to connect
to an embedded server.