Launcher: Implement more robust PATH environment variable handling
This has bothered me for a while- environment variable tweaking is
string manipulation, C APIs are really annoying to do string
manipulation with, and Win32's APIs are even worse thanks to the
char/wchar separation. At the very least, we can isolate it to a single
function, that way the stack-allocated buffer gets automatically
cleaned up when we're done and if I ever have to implement reallocation
I only have one function to touch.